Extract hasPath from common
